R6 class for a fuzzy concept with sparse internal representation

This class implements the data structure and methods for fuzzy concepts.

Methods

Public methods

SparseConcept$new()

SparseConcept$get_extent()

SparseConcept$get_intent()

SparseConcept$print()

SparseConcept$to_latex()

SparseConcept$clone()

Method new()

Creator for objects of class SparseConcept

Usage

SparseConcept$new(extent, intent)

Arguments

extent

(SparseSet) The extent of the concept.

intent

(SparseSet) The intent of the concept.

Returns

An object of class SparseConcept.

Method get_extent()

Internal SparseSet for the extent

Usage

SparseConcept$get_extent()

Returns

The SparseSet representation of the extent.

Method get_intent()

Internal SparseSet for the intent

Usage

SparseConcept$get_intent()

Returns

The SparseSet representation of the intent.

Method print()

Prints the concept to console

Usage

SparseConcept$print()

Returns

A string with the elements of the set and their grades between brackets .

Method to_latex()

Write the concept in LaTeX format

Usage

SparseConcept$to_latex()

Returns

The fuzzy concept in LaTeX.

Method clone()

The objects of this class are cloneable with this method.

Usage

SparseConcept$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Build a formal context and find its concepts
fc_planets <- FormalContext$new(planets)
fc_planets$find_concepts()

# Print the first three concepts
fc_planets$concepts[1:3]

#> ({Mercury, Venus, Earth, Mars, Jupiter, Saturn, Uranus, Neptune, Pluto}, {})
#> ({Earth, Mars, Jupiter, Saturn, Uranus, Neptune, Pluto}, {moon})
#> ({Jupiter, Saturn, Uranus, Neptune, Pluto}, {far, moon})

# Select the first concept:
C <- fc_planets$concepts[1][[1]]

# Get its extent and intent
C$get_extent()

#> {Mercury, Venus, Earth, Mars, Jupiter, Saturn, Uranus, Neptune, Pluto}

C$get_intent()

#> {}