R6 class for a formal context

This class implements the data structure and methods for formal contexts.

Public fields

I: the table for the formal context.

attributes: name of the attributes in the formal context.

objects: name of the objects in the context.

grades_set: set of grades (in 0, 1) of the attributes.

concepts: list of concepts (extent, intent).

implications: extracted implications as an ImplicationSet.

concept_support: vector with the support of the concepts.

implications_support: vector with support of the extracted implications.

Methods

Public methods

FormalContext$new()

FormalContext$add_implications()

FormalContext$compute_concepts()

FormalContext$extract_implications_concepts()

FormalContext$convert_to_transactions()

FormalContext$export_implications_to_arules()

FormalContext$dim()

FormalContext$print()

FormalContext$plot_lattice()

FormalContext$plot_context()

FormalContext$get_concept_support()

FormalContext$get_implication_support()

Method new()

Creator for the Fomal Context class

Usage

FormalContext$new(
  I,
  grades_set = sort(unique(as.vector(I))),
  remove_const = FALSE
)

Arguments

I

(numeric matrix) The table of the formal context.

grades_set

(numeric vector, optional) the anumeration of the grades of the attributes.

remove_const

(logical) If TRUE, remove constant columns. The default is FALSE.

Details

Columns of I must be named, and are the names of the attributes of the formal context.

Returns

An object of the FormalContext class.

Method add_implications()

Add a precomputed implication set

Usage

FormalContext$add_implications(impl_set)

Arguments

impl_set

(ImplicationSet object) The implications to add to this formal context.

Returns

Nothing, just updates the internal implications field.

Method compute_concepts()

Use Ganter Algorithm to compute concepts

Usage

FormalContext$compute_concepts(verbose = FALSE)

Arguments

verbose

(logical) TRUE will provide a verbose output.

Returns

A list with all the concepts in the formal context.

Method extract_implications_concepts()

Use modified Ganter algorithm to compute both concepts and implications

Usage

FormalContext$extract_implications_concepts(verbose = FALSE)

Arguments

verbose

(logical) TRUE will provide a verbose output.

Returns

Nothing, just updates the internal fields concepts and implications.

Method convert_to_transactions()

Convert the formal context to object of class transactions from the arules package

Usage

FormalContext$convert_to_transactions()

Returns

A transactions object.

Method export_implications_to_arules()

Export implications to arules format

Usage

FormalContext$export_implications_to_arules(quality = TRUE)

Arguments

quality

(logical) Compute the interest measures for each rule?

Returns

A rules object from arules package. If quality == TRUE, the function interestMeasure from arules is used.

Method dim()

Dimensions of the formal context

Usage

FormalContext$dim()

Returns

A vector with (number of objects, number of attributes).

Method print()

Prints the formal context

Usage

FormalContext$print()

Returns

Prints information regarding the formal context.

Method plot_lattice()

Plot the concept lattice

Usage

FormalContext$plot_lattice(minsupp = 0, object_names = TRUE)

Arguments

minsupp

(numeric) Only plot the sublattice generated with those concepts whose support is greater than or equal to minsupp.

object_names

(logical) If TRUE, plot object names, otherwise omit them from the diagram.

Returns

Nothing, just plots the graph of the concept lattice.

Method plot_context()

Plot the formal context table

Usage

FormalContext$plot_context()

Returns

Nothing, just plots the formal context.

Method get_concept_support()

Get support of each concept

Usage

FormalContext$get_concept_support()

Returns

A vector with the support of each concept.

Method get_implication_support()

Compute support of each implication

Usage

FormalContext$get_implication_support()

Returns

A vector with the support of each implication