R6 Class for Set of implications

This class implements the structure needed to store implications and the methods associated.

Methods

Public methods

ImplicationSet$new()

ImplicationSet$get_attributes()

ImplicationSet$from_arules()

ImplicationSet$to_arules()

ImplicationSet$add_implication()

ImplicationSet$append_implications()

ImplicationSet$cardinality()

ImplicationSet$is_empty()

ImplicationSet$size()

ImplicationSet$compute_closure()

ImplicationSet$recommend()

ImplicationSet$apply_rules()

ImplicationSet$print()

ImplicationSet$to_latex()

ImplicationSet$get_LHS_matrix()

ImplicationSet$get_RHS_matrix()

ImplicationSet$filter()

ImplicationSet$get_rules()

ImplicationSet$remove_rules()

Method new()

Initialize with an optional name

Usage

ImplicationSet$new(name = "", attributes = c(), lhs = NULL, rhs = NULL)

Arguments

name

(character string) Optional name of the implication set.

attributes

(character vector) Vector of names of the attributes on which we define the implications.

lhs

(a dgCMatrix) Optional initial LHS of the implications stored.

rhs

(a dgCMatrix) Optional initial RHS of the implications stored.

Returns

A new ImplicationSet object.

Method get_attributes()

Get the names of the attributes

Usage

ImplicationSet$get_attributes()

Returns

A character vector with the names of the attributes used in the implications.

Method from_arules()

Import from arules object

Usage

ImplicationSet$from_arules(arules_imp)

Arguments

arules_imp

(rules object) The implications obtained with arules to be imported.

Returns

Nothing, just updates the internal matrices for LHS and RHS.

Method to_arules()

Convert to arules format

Usage

ImplicationSet$to_arules()

Returns

A rules object as used by package arules.

Method add_implication()

Add new implication to the set

Usage

ImplicationSet$add_implication(lhs, rhs)

Arguments

lhs

(a sparse matrix) The LHS of the rule to add.

rhs

(a sparse matrix) The RHS of the rule to add.

Returns

Nothing, just updates the internal matrices for LHS and RHS.

Method append_implications()

Append implications to the current ones.

Usage

ImplicationSet$append_implications(implications)

Arguments

implications

(ImplicationSet) The set of implications to append at the end of the current set.

Returns

Nothing, just updates the internal matrices for LHS and RHS.

Method cardinality()

Cardinality: Number of implications in the set

Usage

ImplicationSet$cardinality()

Returns

The cardinality of the implication set.

Method is_empty()

Empty set

Usage

ImplicationSet$is_empty()

Returns

TRUE if the set of implications is empty, FALSE otherwise.

Method size()

Size: number of attributes in each of LHS and RHS

Usage

ImplicationSet$size()

Returns

A vector with two components: the number of attributes present in each of the LHS and RHS of each implication in the set.

Method compute_closure()

Compute the semantic closure of a fuzzy set with respect to the implication set

Usage

ImplicationSet$compute_closure(S, reduce = FALSE, verbose = FALSE)

Arguments

S

(a SparseSet object) Fuzzy set to compute its closure. Use class SparseSet to build it.

reduce

(logical) Reduce the implications using simplification logic?

verbose

(logical) Show verbose output?

Returns

If reduce == FALSE, the output is a fuzzy set corresponding to the closure of S. If reduce == TRUE, a list with two components: closure, with the closure as above, and implications, the reduced set of implications.

Method recommend()

Generate a recommendation for a subset of the attributes

Usage

ImplicationSet$recommend(S, attribute_filter)

Arguments

S

(a vector) Vector with the grades of each attribute (a fuzzy set).

attribute_filter

(character vector) Names of the attributes to get recommendation for.

Returns

A fuzzy set describing the values of the attributes in attribute_filter within the closure of S.

Method apply_rules()

Apply rules to remove redundancies

Usage

ImplicationSet$apply_rules(
  rules = c("composition", "generalization"),
  batch_size = 25000L,
  parallelize = TRUE,
  reorder = FALSE
)

Arguments

rules

(character vector) Names of the rules to use. See details.

batch_size

(integer) If the number of rules is large, apply the rules by batches of this size.

parallelize

(logical) If possible, should we parallelize the computation among different batches?

reorder

(logical) Should the rules be randonly reordered previous to the computation?

Details

Currently, the implemented rules are "generalization", "simplification", "reduction" and "composition".

Returns

Nothing, just updates the internal matrices for LHS and RHS.

Method print()

Print all implications to text

Usage

ImplicationSet$print()

Returns

A string with all the implications in the set.

Method to_latex()

Export to LaTeX

Usage

ImplicationSet$to_latex(ncols = 1)

Arguments

ncols

(integer) Number of columns for the output.

Returns

A string in LaTeX format that prints nicely all the implications.

Method get_LHS_matrix()

Get internal LHS matrix

Usage

ImplicationSet$get_LHS_matrix()

Returns

A sparse matrix representing the LHS of the implications in the set.

Method get_RHS_matrix()

Get internal RHS matrix

Usage

ImplicationSet$get_RHS_matrix()

Returns

A sparse matrix representing the RHS of the implications in the set.

Method filter()

Filter implications by attributes in LHS and RHS

Usage

ImplicationSet$filter(lhs = NULL, rhs = NULL, drop = FALSE)

Arguments

lhs

(character vector) Names of the attributes to filter the LHS by. If NULL, no filtering is done on the LHS.

rhs

(character vector) Names of the attributes to filter the RHS by. If NULL, no filtering is done on the RHS.

drop

(logical) Remove the rest of attributes in RHS?

Returns

An ImplicationSet that is a subset of the current set, only with those rules which has the attributes in lhs and rhs in their LHS and RHS, respectively.

Method get_rules()

Get a subset of the implication set

Usage

ImplicationSet$get_rules(idx)

Arguments

idx

(integer vector) Indices of the implications to extract.

Returns

A new ImplicationSet with only the rules given by the idx indices.

Method remove_rules()

Remove implications from the set

Usage

ImplicationSet$remove_rules(idx)

Arguments

idx

(integer vector) Indices of implications to remove.

Returns

An ImplicationSet with the implications at given indices removed with respect to the original set.