Check if Set or FormalContext respects an ImplicationSet

Check if Set or FormalContext respects an ImplicationSet

set %respects% imps

Arguments

set
(list of Sets, or a FormalContext) The sets of attributes to check whether they respect the ImplicationSet.

imps
(ImplicationSet) The set of implications to check.

Value

A logical matrix with as many rows as Sets and as many columns as implications in the ImplicationSet. A TRUE in element (i, j) of the result means that the i-th Set respects the j-th implication of the ImplicationSet.

Examples

fc <- FormalContext$new(planets)
fc$find_implications()
imps <- fc$implications$clone()
fc %respects% imps

#> 9 x 10 Matrix of class "lgeMatrix"
#>       imp_01 imp_02 imp_03 imp_04 imp_05 imp_06 imp_07 imp_08 imp_09 imp_10
#> set_1   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_2   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_3   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_4   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_5   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_6   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_7   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_8   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE
#> set_9   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE   TRUE